rebasing

+++ src/main/java/faang/school/accountservice/dto/balance/BalanceDto.java
@@ -2,13 +2,12 @@ package faang.school.accountservice.dto.balance;
 
 import lombok.Builder;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
 public record BalanceDto(
         Long accountId,
-        BigDecimal actualBalance,
+        Double actualBalance,
         LocalDateTime createdAt
 ) {
 }
